+++ domains/graph/nuxt.config.ts
@@ -1,0 +1,4 @@
+// https://nuxt.com/docs/api/configuration/nuxt-config
+export default defineNuxtConfig({
+  routeRules: {},
+})
