+++ domains/rpc/composables/useProfileAssets.ts
