+++ queries/profileFragment.gql
@@ -1,0 +1,41 @@
+fragment profile on Profile {
+  fullName
+  description
+  name
+  id
+  standard
+  interfaces
+  blockNumber
+  links {
+    title
+    url
+  }
+  # tags
+  backgroundImages {
+    url
+    height
+    width
+  }
+  profileImages {
+    width
+    url
+    height
+  }
+  lsp5ReceivedAssets {
+    id
+  }
+  lsp12IssuedAssets {
+    id
+  }
+  ...avatars
+  holds {
+    asset {
+      id
+    }
+    token {
+      baseAsset {
+        id
+      }
+    }
+  }
+}
