+++ queries/profileQuery.gql
@@ -1,0 +1,5 @@
+query profile($id: String!) {
+  Profile_by_pk(id: $id) {
+    ...profile
+  }
+}
